Previously, alignment had to be implemented by each actor, so it was only ever supported by Sprite and BitmapText.  Handle it globally, using m_size.  This way, an actor only needs to set m_size to support alignment.
This means alignment can be set for non-drawing ActorFrames, if a size
is set explicitly.  This allows stretching from the border of a frame
without playing games with extra nested frames.  "setsize,100,100;valign,4;zoomx,.5"

This adds an extra matrix multiply.  This will only happen for actors
actually set to non-centered alignment, which is the exception.

BitmapText still overrides this behavior a bit.  Alignment for text
does two things: changes the alignment of the bounding box vs. the
X/Y position (the generic behavior), and changes the alignment of
the text within the bounding box.


git-svn-id: https://svn.stepmania.com/svn/trunk/stepmania@25357 822b03c2-8f45-0410-a444-95660a741a6f