When an option on a SELECT_ONE row is chosen, don't change focus to "exit".
Only a couple rows are SELECT_ONE; the rest are SELECT_MULTIPLE.
Selecting "exit" automatically changes the flow of the menu on these rows
to be unlike the majority, which makes them clumsy to use: you press start,
then start tapping right to go to "exit" before realizing it's already put you
there (and now you've moved off it).  This makes the menus consistent,
which makes the menu much easier to manipulate intuitively.


git-svn-id: https://svn.stepmania.com/svn/trunk@25833 822b03c2-8f45-0410-a444-95660a741a6f